+++ src/event_timer.c
@@ -78,10 +78,10 @@ static void print_debug_info(struct server *restrict s, const ev_tstamp now)
 		.tcp_out = s->stats.tcp_out - last_stats.tcp_out,
 	};
 	double udp_up, udp_down, tcp_up, tcp_down;
-	udp_up = (dstats.udp_out >> 10u) / dt;
-	udp_down = (dstats.udp_in >> 10u) / dt;
-	tcp_up = (dstats.tcp_in >> 10u) / dt;
-	tcp_down = (dstats.tcp_out >> 10u) / dt;
+	udp_up = (double)(dstats.udp_out >> 10u) / dt;
+	udp_down = (double)(dstats.udp_in >> 10u) / dt;
+	tcp_up = (double)(dstats.tcp_in >> 10u) / dt;
+	tcp_down = (double)(dstats.tcp_out >> 10u) / dt;
 	LOGD_F("traffic(KiB/s) udp up/down: %.1f/%.1f; tcp up/down: %.1f/%.1f; efficiency: %.1f%%/%.1f%%",
 	       udp_up, udp_down, tcp_up, tcp_down, tcp_up / udp_up * 100.0,
 	       tcp_down / udp_down * 100.0);
